+++ dev/user.clj
@@ -1,42 +1,11 @@
 (ns user
   (:require
    [clj-reload.core :as reload]
-   [environ.core :refer [env]]
-   [org.httpkit.server :as hk]
-   [ring.middleware.reload :refer [wrap-reload]]
-   [taoensso.telemere :as t]
-   [hkimjp.jpy.routes :refer [root-handler]]
-   [hkimjp.datascript :as ds]))
+   [taoensso.telemere :as tel]
+   [hkimjp.jpy.system :refer [start-system stop-system]]))
 
 ;--------------------------
-; ring-devel
-(defonce server (atom nil))
-
-(defn start-server []
-  (let [handler (wrap-reload #'root-handler)]
-    (let [port (parse-long (or (env :port) "3000"))]
-      (reset! server (hk/run-server handler {:port port}))
-      (println (str "http-kit started at port " port)))))
-
-(defn stop-server []
-  (@server)
-  (reset! server nil))
-
-(defn start-system []
-  (t/log! {:level :info
-           :id "start-system"
-           :msg (env :develop)
-           :data {:datascript (env :datascript)}})
-  (try
-    (ds/start-or-restore {:url (env :datascript)})
-    (start-server)
-    (catch Exception e
-      (t/log! :fatal (.getMessage e))
-      (System/exit 0))))
-
-(defn stop-system []
-  (stop-server)
-  (ds/stop))
+(tel/set-min-level! :debug)
 
 (start-system)
 ; (stop-system)
@@ -57,3 +26,4 @@
 
 ; (reload/reload)
 
+;-------------------